Fix: req 17 and other reqs deads
